// typings/BizzomateIFrameProps.d.ts
/**
 * This file was generated from BizzomateIFrame.xml
 * WARNING: All changes made to this file will be overwritten
 * @author Mendix UI Content Team
 */
import { CSSProperties } from "react";
import { DynamicValue } from "mendix";

export interface BizzomateIFrameContainerProps {
    name: string;
    class: string;
    style?: CSSProperties;
    tabIndex?: number;
    sourceURL?: DynamicValue<string>;
    sourceDoc?: DynamicValue<string>;
    title?: DynamicValue<string>;
    width: string;
    height: string;
    allow: string;
    referrerPolicy: string;
    sandbox: string;
    loading: string;
}

export interface BizzomateIFramePreviewProps {
    className: string;
    style: string;
    styleObject?: CSSProperties;
    readOnly: boolean;
    sourceURL: string;
    sourceDoc: string;
    title: string;
    width: string;
    height: string;
    allow: string;
    referrerPolicy: string;
    sandbox: string;
    loading: string;
}